+++ middleware/auth.global.ts
@@ -1,4 +1,4 @@
-export default defineNuxtRouteMiddleware(async () => {
+export default defineNuxtRouteMiddleware(async (to, from) => {
   // Below condition process only on client side
 
 const userToken = localStorage.getItem('token')
@@ -7,4 +7,10 @@ const userStore = useUserStore()
 const {token} = storeToRefs(userStore);
   if(userToken)
      token.value = userToken;
+
+ if(to.fullPath.includes('/admin') && !token.value) {
+   return navigateTo('/log-in')
+}
+
+
 });
