+++ pages/index.vue
@@ -1,103 +1,69 @@
 <template>
-    <div class="bg-white my-12">
-      <div v-if="!filteredProducts.length" class="flex flex-col items-center justify-center h-[50vh]">
-      <div class="text-gray-500 tracking-wide space-y-5">
-        <div class="flex justify-center items-center space-x-4">
-          <span
-            class="block animate-spin border-t-2 h-20 w-20 border-l-2 border-b-2 border-r-2 border-t-primary-light rounded-full border-gray-300"
-            
-          />
-        </div>
-        <div class="flex flex-col items-center animate-pulse">
-          <p>Please wait.</p>
-          <p>It'll just take a moment.</p>
-        </div>
-      </div>
-    </div>
-      <div v-else class="max-w-7xl mx-auto overflow-hidden sm:px-6 lg:px-8">
-        <div class="flex justify-between my-2">
-          <h2 class="text-3xl font-semibold">Products</h2>
-        </div> 
-        <div class="grid grid-cols-1 gap-4 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 border-l border-t border-gray-200 my-10">
-          <div
-            v-for="product in filteredProducts"
-            :key="product.id"
-            class="p-4 border-r border-b border-gray-200 group relative bg-white rounded-lg shadow-lg hover:shadow-xl transition-shadow"
-          >
-          <NuxtLink :to="`/product/${product.id}`">
-
-            <div class="rounded-lg overflow-hidden h-64">
-              <img
-                :src="product.image"
-                :alt="product.imageAlt"
-                class="w-full h-full object-center object-contain transition-opacity group-hover:opacity-75"
-              >
-            </div>
-            <div class="p-4 text-center">
-              <h3 class="text-lg font-semibold text-gray-900 hover:underline">
-                  {{ product.title.substring(0, 19) }}...
-              </h3>
-              <div class="mt-2">
-                <div class="flex justify-center">
-                  <StarIcon
-                    v-for="rating in [0, 1, 2, 3, 4]"
-                    :key="rating"
-                    :class="[
-                     product?.rating && product?.rating.rate > rating ? 'text-yellow-400' : 'text-gray-200',
-                      'h-5 w-5'
-                    ]"
-                  />
+    <div class="bg-white">
+      <main>
+        <div class="relative isolate">
+          <svg class="absolute inset-x-0 top-0 -z-10 h-[64rem] w-full stroke-gray-200 [mask-image:radial-gradient(32rem_32rem_at_center,white,transparent)]" aria-hidden="true">
+            <defs>
+              <pattern id="1f932ae7-37de-4c0a-a8b0-a6e3b4d44b84" width="200" height="200" x="50%" y="-1" patternUnits="userSpaceOnUse">
+                <path d="M.5 200V.5H200" fill="none" />
+              </pattern>
+            </defs>
+            <svg x="50%" y="-1" class="overflow-visible fill-gray-50">
+              <path d="M-200 0h201v201h-201Z M600 0h201v201h-201Z M-400 600h201v201h-201Z M200 800h201v201h-201Z" stroke-width="0" />
+            </svg>
+            <rect width="100%" height="100%" stroke-width="0" fill="url(#1f932ae7-37de-4c0a-a8b0-a6e3b4d44b84)" />
+          </svg>
+          <div class="absolute left-1/2 right-0 top-0 -z-10 -ml-24 transform-gpu overflow-hidden blur-3xl lg:ml-24 xl:ml-48" aria-hidden="true">
+            <div class="aspect-[801/1036] w-[50.0625rem] bg-gradient-to-tr from-[#ff80b5] to-[#9089fc] opacity-30" style="clip-path: polygon(63.1% 29.5%, 100% 17.1%, 76.6% 3%, 48.4% 0%, 44.6% 4.7%, 54.5% 25.3%, 59.8% 49%, 55.2% 57.8%, 44.4% 57.2%, 27.8% 47.9%, 35.1% 81.5%, 0% 97.7%, 39.2% 100%, 35.2% 81.4%, 97.2% 52.8%, 63.1% 29.5%)" />
+          </div>
+          <div class="overflow-hidden">
+            <div class="mx-auto max-w-7xl px-6 pb-32 pt-36 sm:pt-60 lg:px-8 lg:pt-32">
+              <div class="mx-auto max-w-2xl gap-x-14 lg:mx-0 lg:flex lg:max-w-none lg:items-center">
+                <div class="w-full max-w-xl lg:shrink-0 xl:max-w-2xl">
+                  <h1 class="text-4xl font-bold tracking-tight text-gray-900 sm:text-6xl">We’re changing the way people connect.</h1>
+                  <p class="relative mt-6 text-lg leading-8 text-gray-600 sm:max-w-md lg:max-w-none">Cupidatat minim id magna ipsum sint dolor qui. Sunt sit in quis cupidatat mollit aute velit. Et labore commodo nulla aliqua proident mollit ullamco exercitation tempor. Sint aliqua anim nulla sunt mollit id pariatur in voluptate cillum.</p>
+                  <div class="mt-10 flex items-center gap-x-6">
+                    <a href="#" class="rounded-md bg-indigo-600 px-3.5 py-2.5 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600">Get started</a>
+                    <a href="#" class="text-sm font-semibold leading-6 text-gray-900">Live demo <span aria-hidden="true">→</span></a>
+                  </div>
                 </div>
-                <p class="text-sm text-gray-500 mt-1">{{ product.rating?.count }} reviews</p>
+                <div class="mt-14 flex justify-end gap-8 sm:-mt-44 sm:justify-start sm:pl-20 lg:mt-0 lg:pl-0">
+                  <div class="ml-auto w-44 flex-none space-y-8 pt-32 sm:ml-0 sm:pt-80 lg:order-last lg:pt-36 xl:order-none xl:pt-80">
+                    <div class="relative">
+                      <img src="https://images.unsplash.com/photo-1557804506-669a67965ba0?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&h=528&q=80" alt="" class="aspect-[2/3] w-full rounded-xl bg-gray-900/5 object-cover shadow-lg" >
+                      <div class="pointer-events-none absolute inset-0 rounded-xl ring-1 ring-inset ring-gray-900/10" />
+                    </div>
+                  </div>
+                  <div class="mr-auto w-44 flex-none space-y-8 sm:mr-0 sm:pt-52 lg:pt-36">
+                    <div class="relative">
+                      <img src="https://images.unsplash.com/photo-1485217988980-11786ced9454?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&h=528&q=80" alt="" class="aspect-[2/3] w-full rounded-xl bg-gray-900/5 object-cover shadow-lg" >
+                      <div class="pointer-events-none absolute inset-0 rounded-xl ring-1 ring-inset ring-gray-900/10" />
+                    </div>
+                    <div class="relative">
+                      <img src="https://images.unsplash.com/photo-1559136555-9303baea8ebd?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&crop=focalpoint&fp-x=.4&w=396&h=528&q=80" alt="" class="aspect-[2/3] w-full rounded-xl bg-gray-900/5 object-cover shadow-lg" >
+                      <div class="pointer-events-none absolute inset-0 rounded-xl ring-1 ring-inset ring-gray-900/10" />
+                    </div>
+                  </div>
+                  <div class="w-44 flex-none space-y-8 pt-32 sm:pt-0">
+                    <div class="relative">
+                      <img src="https://images.unsplash.com/photo-1670272504528-790c24957dda?ixlib=rb-4.0.3&ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&crop=left&w=400&h=528&q=80" alt="" class="aspect-[2/3] w-full rounded-xl bg-gray-900/5 object-cover shadow-lg" >
+                      <div class="pointer-events-none absolute inset-0 rounded-xl ring-1 ring-inset ring-gray-900/10" />
+                    </div>
+                    <div class="relative">
+                      <img src="https://images.unsplash.com/photo-1670272505284-8faba1c31f7d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDF8MHxwaG90by1wYWdlfHx8fGVufDB8fHx8&auto=format&fit=crop&h=528&q=80" alt="" class="aspect-[2/3] w-full rounded-xl bg-gray-900/5 object-cover shadow-lg" >
+                      <div class="pointer-events-none absolute inset-0 rounded-xl ring-1 ring-inset ring-gray-900/10" />
+                    </div>
+                  </div>
+                </div>
               </div>
-              <p class="text-base font-medium text-gray-900 mt-4">${{ product.price }}</p>
             </div>
-          </NuxtLink>
-
           </div>
         </div>
-      </div>
+      </main>
     </div>
   </template>
   
-<script setup lang="ts">
-import { useProductStore } from "@/stores/useProductStore";
-import { StarIcon } from '@heroicons/vue/24/solid';
-import type { ProductDetail } from '~/types/products';
-
-const productStore = useProductStore();
-const { productList, searchTerm } = storeToRefs(productStore);
-const filteredProducts = ref<ProductDetail[]>([]);
-let debounceTimeout: ReturnType<typeof setTimeout>;
-
-// Function to filter products
-const filterProducts = () => {
-  if (!searchTerm.value) {
-    filteredProducts.value = productList.value;
-  } else {
-    const lowerSearchTerm = searchTerm.value.toLowerCase();
-    filteredProducts.value = productList.value.filter(product =>
-      product.title.toLowerCase().includes(lowerSearchTerm) ||
-      product.category.toLowerCase().includes(lowerSearchTerm)
-    );
-  }
-};
-// Watcher with debouncing
-watch(searchTerm, () => {
-  // Clear the previous timeout if it exists
-  clearTimeout(debounceTimeout);
-
-  // Set a new timeout to delay filtering
-  debounceTimeout = setTimeout(() => {
-    filterProducts();
-  }, 300); // Adjust delay as needed
-});
+  <script setup lang="ts">
 
 
-onMounted(async() => {
- await productStore.fetchProductList();
- filterProducts()
-})
-
-
-</script>
+  </script>
